# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Biblio API",
        "description": "API untuk Biblio",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/genres": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "List semua genre",
                "operationId": "getGenres",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Genre"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/genres/{id}": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "Detail genre",
                "operationId": "getGenre",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessageResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/genres/with-books": {
            "get": {
                "tags": [
                    "Genres"
                ],
                "summary": "List genre beserta preview buku (untuk beranda)",
                "operationId": "getGenreWithBooks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GenreWithBooksResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiMessageResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Book": {
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "8fd93ee3-3815-4334-9218-7ebc50e1343f"
                    },
                    "genre_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "title": {
                        "type": "string",
                        "example": "Judul Buku"
                    },
                    "isbn": {
                        "type": "string",
                        "example": "123ABC"
                    },
                    "description": {
                        "type": "string",
                        "nullable": true
                    },
                    "author": {
                        "type": "string",
                        "example": "Seseorang"
                    },
                    "publisher": {
                        "type": "string",
                        "nullable": true
                    },
                    "lang": {
                        "type": "string",
                        "example": "id"
                    },
                    "published_date": {
                        "type": "string",
                        "nullable": true
                    },
                    "format": {
                        "type": "string",
                        "example": "pdf"
                    },
                    "page_count": {
                        "type": "integer",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "example": 1234
                    },
                    "cover_url": {
                        "type": "string",
                        "nullable": true
                    },
                    "cover_sm": {
                        "type": "string",
                        "nullable": true
                    },
                    "cover_md": {
                        "type": "string",
                        "nullable": true
                    },
                    "cover_lg": {
                        "type": "string",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "genre": {
                        "$ref": "#/components/schemas/Genre"
                    }
                },
                "type": "object"
            },
            "Genre": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Genre Keren"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "ApiMessageResponse": {
                "properties": {
                    "message": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
            "GenreWithBooksResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Biologi"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "updated_at": {
                        "type": "string"
                    },
                    "books": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Book"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}